+++ package.json
@@ -30,5 +30,8 @@
     "prettier": "^2.0.5",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11"
+  },
+  "dependencies": {
+    "@babel/preset-env": "^7.10.2"
   }
 }
