+++ src/app.js
@@ -1,33 +1,6 @@
-const playerController = (() => {
-  class Player {
-    constructor(name, symbol) {
-      this.name = name;
-      this.moves = [];
-      this.symbol = symbol;
-      this.winner = false;
-      this.active = false;
-    }
-  }
+import playerController from './playerController';
 
-  return {
-    addPlayer(name, symbol) {
-      return new Player(name, symbol);
-    },
-    addMove(move) {
-      this.moves.push(move);
-    },
-
-    addWinener() {
-      this.winner = true;
-    },
-
-    activePlayer(active) {
-      this.active = active;
-    },
-  };
-})();
-
-const gameBoardController = ((playerController) => {
+export const gameBoardController = ((playerController) => {
   class GameBoard {
     constructor(xPlayer, oPlayer) {
       this.xPlayer = xPlayer;
@@ -188,5 +161,3 @@ const globalController = ((gameBoardController) => {
 })(gameBoardController);
 
 globalController.init();
-
-export default gameBoardController;
