--- src/app.test.js
@@ -1,8 +1,0 @@
-import gameBoardController from './app.js';
-
-describe('#checkWinningMove', () => {
-  test('winning  combination', () => {
-    gameBoardController.winConditions = ['1', '2', '3', '', '', '', '', '', ''];
-    expect(gameBoardController.checkWinningMove()).toBe(true);
-  });
-});
